Implement Automatic SRM Saving (#539)
SRM files are static RAM files. These are the equivelent of a memory card or battery backed memory storage in console cartridges.

This implements a change where every 60 seconds the emulator will upload the generated SRM file (if one is created) to the Gaseous server for storage. The SRM file is only saved if there are changes.

This SRM file is automatically restored when the emulator is started.